initial scaling/resolution work

+++ core/src/main/java/org/nolat/explosions/Config.java
@@ -3,7 +3,7 @@ package org.nolat.explosions;
 public class Config {
     public static final String NAME = "Explosions";
     public static final String VERSION = "0.1";
-    public static final int WIDTH = 1280;
-    public static final int HEIGHT = 720;
+    public static final float WIDTH = 1280;
+    public static final float HEIGHT = 720;
     public static boolean debug = false;
 }
